-- update win scene --

+++ trunk/Code/cocos2d-2.0-x-2.0.3/ProtectLand/Classes/Resource Management/LevelManager.h
@@ -71,7 +71,7 @@ public:
 	SLevelInfomation* GetListLevel(int& NumOfLevel);
 	bool ResetLevel(const char* pszPath, int iNumOfLevel);
 	bool ResetLevel(const char* pszPath);
-	bool SetLevelInfo(int iMapCurrent, int iLevelCurrent, int iCoin, int iLevelTower, int iDameTowerCurrent, float iTowerSpeed, int iTowerHp, int iRateDoubleDamge);
+	bool SetLevelInfo(int iMapCurrent, int iLevelCurrent, int iCoin, int iLevelTower, int iDameTowerCurrent, int iTowerSpeed, int iTowerHp, int iRateDoubleDamge);
 	
 	
 	bool CreateFileFirstGame(const char* pc_FileName, int iNumOfLevel);
